Remove git hook applying `[not verified]` (#33075)
The hook flagging commits as `[not verified]` when the pre-commit hook
was skipped is a legacy of the days before we had CI running the linting
and blocking merges. These days it mostly serves to annoy people who
rebase instead of merge when updating their PRs.

After some discussion in pdWQjU-v6-p2, we've decided to remove this
flagging.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6191277004